#20 Add Item struct to components package and implement InsertDB function
