+++ backend/components/type.go
@@ -1,0 +1,8 @@
+package components
+
+type Item struct {
+	ID       int    `json:"id"`
+	Category string `json:"category"`
+	Name     string `json:"name"`
+	Weight   int    `json:"weight"`
+}
