+++ backend/test/test_update.rest
@@ -1,0 +1,9 @@
+POST http://localhost:8080/ccc/updateDB HTTP/1.1
+Content-Type: application/json
+Accept: application/json
+
+{
+    "category": "class",
+    "name": "教科書",
+    "weight": 1
+}
